refactor(magic): split models.py into thematic submodules
Convert world/magic/models.py into world/magic/models/ package per
Scope 6 §4.4. Zero behaviour change: __init__.py re-exports the full
public surface so external callers are unaffected. Verified no spurious
migration emitted via makemigrations --dry-run.
